fixed binary assignment bug, refactored file paths for data and main analysis
binary bug was in assigning nearest site for up and down. nearest_site is a 1x1 dataframe, and using nearest_site[1] in the filtering of dw/uw caused wells which were down/up of more than one site to be dropped from this part of the assignment. This was changed to nearest_site$site, which fixed the issue.